fix(l1): exclude obsoleted txs from cumulative-balance sum; move mempool tests to ethrex-test
