+++ crates/blockchain/mempool.rs
@@ -906,13 +906,17 @@ impl Mempool {
     pub fn sum_cost_for_sender(
         &self,
         sender: Address,
+        account_nonce: u64,
         exclude: Option<H256>,
     ) -> Result<U256, MempoolError> {
         let inner = self.read()?;
         let mut total = U256::zero();
+        // Start at `account_nonce`, not 0, so obsoleted txs (nonce below the
+        // sender's on-chain nonce — already executed but not yet pruned from the
+        // pool) don't count toward the sender's required balance.
         for (_key, hash) in inner
             .txs_by_sender_nonce
-            .range((sender, 0)..=(sender, u64::MAX))
+            .range((sender, account_nonce)..=(sender, u64::MAX))
         {
             if Some(*hash) == exclude {
                 continue;
@@ -1189,98 +1193,4 @@ pub fn transaction_intrinsic_gas(
         0
     };
     Ok(intrinsic.max(calldata_floor))
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use ethrex_common::types::{EIP1559Transaction, TxKind};
-
-    const MEMPOOL_MAX_SIZE_TEST: usize = 10_000;
-
-    fn build_tx(nonce: u64, max_fee_per_gas: u64, gas_limit: u64, value: U256) -> Transaction {
-        Transaction::EIP1559Transaction(EIP1559Transaction {
-            nonce,
-            max_priority_fee_per_gas: 0,
-            max_fee_per_gas,
-            gas_limit,
-            to: TxKind::Call(Address::from_low_u64_be(1)),
-            value,
-            ..Default::default()
-        })
-    }
-
-    fn insert_tx(mempool: &Mempool, sender: Address, tx: Transaction) -> H256 {
-        let hash = H256::random();
-        mempool
-            .add_transaction(hash, sender, MempoolTransaction::new(tx, sender), None)
-            .expect("add_transaction");
-        hash
-    }
-
-    #[test]
-    fn sum_cost_for_sender_empty_pool_is_zero() {
-        let mempool = Mempool::new(MEMPOOL_MAX_SIZE_TEST);
-        let sender = Address::random();
-
-        let total = mempool.sum_cost_for_sender(sender, None).expect("sum");
-        assert_eq!(total, U256::zero());
-    }
-
-    #[test]
-    fn sum_cost_for_sender_sums_multiple_nonces() {
-        let mempool = Mempool::new(MEMPOOL_MAX_SIZE_TEST);
-        let sender = Address::random();
-
-        let tx0 = build_tx(0, 10, 21_000, U256::from(100u64));
-        let tx1 = build_tx(1, 20, 21_000, U256::from(200u64));
-        let tx2 = build_tx(2, 30, 21_000, U256::from(300u64));
-
-        let expected = tx0.cost_without_base_fee().unwrap()
-            + tx1.cost_without_base_fee().unwrap()
-            + tx2.cost_without_base_fee().unwrap();
-
-        insert_tx(&mempool, sender, tx0);
-        insert_tx(&mempool, sender, tx1);
-        insert_tx(&mempool, sender, tx2);
-
-        let total = mempool.sum_cost_for_sender(sender, None).expect("sum");
-        assert_eq!(total, expected);
-    }
-
-    #[test]
-    fn sum_cost_for_sender_ignores_other_senders() {
-        let mempool = Mempool::new(MEMPOOL_MAX_SIZE_TEST);
-        let sender_a = Address::random();
-        let sender_b = Address::random();
-
-        let tx_a = build_tx(0, 10, 21_000, U256::from(100u64));
-        let tx_b = build_tx(0, 50, 21_000, U256::from(999u64));
-
-        let expected_a = tx_a.cost_without_base_fee().unwrap();
-
-        insert_tx(&mempool, sender_a, tx_a);
-        insert_tx(&mempool, sender_b, tx_b);
-
-        let total_a = mempool.sum_cost_for_sender(sender_a, None).expect("sum a");
-        assert_eq!(total_a, expected_a);
-    }
-
-    #[test]
-    fn sum_cost_for_sender_after_remove_drops_that_tx() {
-        let mempool = Mempool::new(MEMPOOL_MAX_SIZE_TEST);
-        let sender = Address::random();
-
-        let tx0 = build_tx(0, 10, 21_000, U256::from(100u64));
-        let tx1 = build_tx(1, 10, 21_000, U256::from(200u64));
-        let expected_after = tx1.cost_without_base_fee().unwrap();
-
-        let hash0 = insert_tx(&mempool, sender, tx0);
-        insert_tx(&mempool, sender, tx1);
-
-        mempool.remove_transaction(&hash0).expect("remove");
-
-        let total = mempool.sum_cost_for_sender(sender, None).expect("sum");
-        assert_eq!(total, expected_after);
-    }
 }
